Add config module and cleanup
