+++ editor.c
@@ -9,6 +9,7 @@
 #include <wctype.h>
 #include <time.h>
 #include <signal.h>
+#include "config.h"
 #include <nl_types.h>
 #define CONTROL_KEYS 1
 #define COMMANDS 2
@@ -42,21 +43,13 @@ extern int scr_vert;
 extern int scr_horz;
 extern int absolute_lin;
 extern int edit;
-extern int info_window;
-extern int nohighlight;
 extern int clear_com_win;
-extern int emacs_keys_mode;
-extern int right_margin;
 extern int horiz_offset;
 extern int gold;
 extern int case_sen;
 extern int shell_fork;
 extern int in;
 extern char *in_file_name;
-extern char *no_file_string;
-extern char *com_win_message;
-extern char *separator;
-extern char *ree_no_file_msg;
 extern WINDOW *com_win;
 extern WINDOW *text_win;
 extern WINDOW *info_win;
@@ -72,9 +65,6 @@ extern int last_col;
 extern int local_LINES;
 extern int local_COLS;
 extern int info_type;
-extern int eightbit;
-extern int ee_chinese;
-extern char *print_command;
 extern char *mode_strings[];
 extern struct menu_entries modes_menu[];
 extern struct menu_entries leave_menu[];
@@ -83,114 +73,6 @@ extern struct menu_entries search_menu[];
 extern struct menu_entries spell_menu[];
 extern struct menu_entries misc_menu[];
 extern struct menu_entries main_menu[];
-extern char *help_text[];
-extern char *emacs_help_text[];
-extern char *control_keys[];
-extern char *emacs_control_keys[];
-extern char *command_strings[];
-extern char *ascii_code_str;
-extern char *printer_msg_str;
-extern char *command_str;
-extern char *file_write_prompt_str;
-extern char *file_read_prompt_str;
-extern char *char_str;
-extern char *unkn_cmd_str;
-extern char *non_unique_cmd_msg;
-extern char *line_num_str;
-extern char *line_len_str;
-extern char *current_file_str;
-extern char *usage0;
-extern char *usage1;
-extern char *usage2;
-extern char *usage3;
-extern char *usage4;
-extern char *file_is_dir_msg;
-extern char *new_file_msg;
-extern char *cant_open_msg;
-extern char *open_file_msg;
-extern char *file_read_fin_msg;
-extern char *reading_file_msg;
-extern char *read_only_msg;
-extern char *file_read_lines_msg;
-extern char *save_file_name_prompt;
-extern char *file_not_saved_msg;
-extern char *changes_made_prompt;
-extern char *yes_char;
-extern char *file_exists_prompt;
-extern char *create_file_fail_msg;
-extern char *writing_file_msg;
-extern char *file_written_msg;
-extern char *searching_msg;
-extern char *str_not_found_msg;
-extern char *search_prompt_str;
-extern char *exec_err_msg;
-extern char *continue_msg;
-extern char *menu_cancel_msg;
-extern char *menu_size_err_msg;
-extern char *press_any_key_msg;
-extern char *shell_prompt;
-extern char *formatting_msg;
-extern char *shell_echo_msg;
-extern char *spell_in_prog_msg;
-extern char *margin_prompt;
-extern char *restricted_msg;
-extern char *ON;
-extern char *OFF;
-extern char *HELP;
-extern char *WRITE;
-extern char *READ;
-extern char *LINE;
-extern char *FILE_str;
-extern char *CHARACTER;
-extern char *REDRAW;
-extern char *RESEQUENCE;
-extern char *AUTHOR;
-extern char *VERSION;
-extern char *CASE;
-extern char *NOCASE;
-extern char *EXPAND;
-extern char *NOEXPAND;
-extern char *Exit_string;
-extern char *QUIT_string;
-extern char *INFO;
-extern char *NOINFO;
-extern char *MARGINS;
-extern char *NOMARGINS;
-extern char *AUTOFORMAT;
-extern char *NOAUTOFORMAT;
-extern char *Echo;
-extern char *PRINTCOMMAND;
-extern char *RIGHTMARGIN;
-extern char *HIGHLIGHT;
-extern char *NOHIGHLIGHT;
-extern char *EIGHTBIT;
-extern char *NOEIGHTBIT;
-extern char *EMACS_string;
-extern char *NOEMACS_string;
-extern char *conf_dump_err_msg;
-extern char *conf_dump_success_msg;
-extern struct menu_entries config_dump_menu[];
-extern char *conf_not_saved_msg;
-extern char *menu_too_lrg_msg;
-extern char *more_above_str;
-extern char *more_below_str;
-extern char *chinese_cmd;
-extern char *nochinese_cmd;
-extern char *commands[];
-extern char *init_strings[];
-extern nl_catd catalog;
-extern char *init_name[];
-extern int expand_tabs;
-extern int observ_margins;
-extern int auto_format;
-
-extern int unique_test(char *string, char *list[]);
-extern int compare(char *string1, char *string2, int sensitive);
-extern char *next_ascii_word(char *string);
-extern void echo_string(char *string);
-extern char *catgetlocal(int, const char *);
-extern char *get_string(const char *prompt, int advance);
-extern void command(char *cmd_str1);
 
 /* external helpers still in ee.c */
 extern void get_options(int numargs, char *arguments[]);
